Move estimate prioritization tests to its own file
