+++ app/utils/estimate_prioritization_new/__init__.py
@@ -1,2 +1,3 @@
 from .pooling_functions import pooling_function_maps
 from .prioritization_criteria import prioritization_criteria_testadj, prioritization_criteria_testunadj
+from .estimate_prioritization import get_prioritized_estimates
